Restructure modules and provide browser bundle
