+++ datum/index.js
@@ -1,0 +1,2 @@
+exports.bd09 = require('./bd-09');
+exports.gcj02 = require('./gcj-02');
