+++ examples/transforms.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>projzh transforms</title>
+  </head>
+  <body>
+    Check the console for output.
+    <script src="../index.min.js"></script>
+    <script src="./transforms.js"></script>
+  </body>
+</html>
