+++ projection/index.js
@@ -1,0 +1,2 @@
+exports.baiduMercator = require('./baidu-mercator');
+exports.sphericalMercator = require('./spherical-mercator');
